Move metadata to a separate module
